Allow custom HTTP status for throttled requests
