+++ concurrency-limits-servlet/src/test/java/com/netflix/concurrency/limits/ConcurrencyLimitServletFilterTest.java
@@ -1,88 +1,86 @@
 package com.netflix.concurrency.limits;
 
-import com.netflix.concurrency.limits.executors.BlockingAdaptiveExecutor;
-import com.netflix.concurrency.limits.limit.FixedLimit;
-import com.netflix.concurrency.limits.limit.VegasLimit;
-import com.netflix.concurrency.limits.limiter.SimpleLimiter;
+
 import com.netflix.concurrency.limits.servlet.ConcurrencyLimitServletFilter;
-import com.netflix.concurrency.limits.servlet.ServletLimiterBuilder;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.Mock;
+import org.mockito.Mockito;
+import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.mock.web.MockFilterChain;
+import org.springframework.mock.web.MockHttpServletRequest;
+import org.springframework.mock.web.MockHttpServletResponse;
+
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServletRequest;
 
 import java.io.IOException;
-import java.security.Principal;
-import java.util.EnumSet;
-import java.util.concurrent.Executors;
-import java.util.concurrent.RejectedExecutionException;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.Optional;
 
-import javax.servlet.DispatcherType;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNull;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
 
-import org.eclipse.jetty.servlet.FilterHolder;
-import org.junit.ClassRule;
-import org.junit.Ignore;
-import org.junit.Test;
-
+@ExtendWith(MockitoExtension.class)
 public class ConcurrencyLimitServletFilterTest {
-    @ClassRule
-    public static HttpServerRule server = new HttpServerRule(context -> {
-        context.addServlet(HelloServlet.class, "/");
-        
-        Limiter<HttpServletRequest> limiter = new ServletLimiterBuilder()
-                .limit(FixedLimit.of(10))
-                .partitionByUserPrincipal(Principal::getName)
-                .partition("live", 0.8)
-                .partition("batch", 0.2)
-                .build();
 
-        FilterHolder holder = new FilterHolder();
-        holder.setFilter(new ConcurrencyLimitServletFilter(limiter));
-        
-        context.addFilter(holder, "/*", EnumSet.of(DispatcherType.REQUEST));
-    });
-    
+    @Mock
+    Limiter<HttpServletRequest> limiter;
+
+    @Mock
+    Limiter.Listener listener;
+
     @Test
-    @Ignore
-    public void simulation() throws Exception {
-        Limit limit = VegasLimit.newDefault();
-        BlockingAdaptiveExecutor executor = new BlockingAdaptiveExecutor(
-                SimpleLimiter.newBuilder().limit(limit).build());
-        AtomicInteger errors = new AtomicInteger();
-        AtomicInteger success = new AtomicInteger();
-        
-        Executors.newSingleThreadScheduledExecutor().scheduleAtFixedRate(() -> {
-            System.out.println(String.format("errors=%d success=%d limit=%s", errors.getAndSet(0), success.getAndSet(0), limit));
-        }, 1, 1, TimeUnit.SECONDS);
-        
-        
-        while (true) {
-            executor.execute(() -> {
-                try {
-                    server.get("/batch");
-                    success.incrementAndGet();
-                } catch (Exception e) {
-                    errors.incrementAndGet();
-                    throw new RejectedExecutionException();
-                }
-            });
-        }
-    }
-    
-    public static class HelloServlet extends HttpServlet {
-        private static final long serialVersionUID = 1L;
+    public void testDoFilterAllowed() throws ServletException, IOException {
 
-        @Override
-        protected void doGet( HttpServletRequest request, HttpServletResponse response ) throws ServletException, IOException {
-            try {
-                TimeUnit.MILLISECONDS.sleep(100);
-            } catch (InterruptedException e) {
-            }
-            response.setContentType("text/html");
-            response.setStatus(HttpServletResponse.SC_OK);
-            response.getWriter().println("Hello from HelloServlet");
-        }
+        ConcurrencyLimitServletFilter filter = new ConcurrencyLimitServletFilter(limiter);
+
+        when(limiter.acquire(any())).thenReturn(Optional.of(listener));
+
+        MockHttpServletRequest request = new MockHttpServletRequest();
+        MockHttpServletResponse response = new MockHttpServletResponse();
+        MockFilterChain filterChain = new MockFilterChain();
+
+
+        filter.doFilter(request, response, filterChain);
+        assertEquals(request, filterChain.getRequest(), "Request should be passed to the downstream chain");
+        assertEquals(response, filterChain.getResponse(), "Response should be passed to the downstream chain");
+
+        verify(listener).onSuccess();
+    }
+
+    @Test
+    public void testDoFilterThrottled() throws ServletException, IOException {
+        ConcurrencyLimitServletFilter filter = new ConcurrencyLimitServletFilter(limiter);
+
+        //Empty means to throttle this request
+        when(limiter.acquire(any())).thenReturn(Optional.empty());
+
+        MockHttpServletResponse response = new MockHttpServletResponse();
+        MockFilterChain filterChain = new MockFilterChain();
+
+
+        filter.doFilter(new MockHttpServletRequest(), response, filterChain);
+
+        assertNull(filterChain.getRequest(), "doFilter should not be called on the filterchain");
+
+        assertEquals(429, response.getStatus(), "Status should be 429 - too many requests");
+    }
+
+    @Test
+    public void testDoFilterThrottledCustomStatus() throws ServletException, IOException {
+        final int customThrottleStatus = 503;
+        ConcurrencyLimitServletFilter filter = new ConcurrencyLimitServletFilter(limiter, customThrottleStatus);
+
+        //Empty means to throttle this request
+        when(limiter.acquire(any())).thenReturn(Optional.empty());
+
+        MockHttpServletResponse response = new MockHttpServletResponse();
+
+        filter.doFilter(new MockHttpServletRequest(), response,  new MockFilterChain());
+
+        assertEquals(customThrottleStatus, response.getStatus(), "custom status should be respected");
     }
 }
